Added logging and a missing component.
