+++ packages/spikes/gulf_client/lib/gulf_client.dart
@@ -8,10 +8,11 @@
 /// from a JSON-based definition provided by a server.
 library;
 
+export 'src/core/gulf_agent_connector.dart';
 export 'src/core/interpreter.dart';
 export 'src/core/widget_registry.dart';
-export 'src/core/gulf_agent_connector.dart';
 export 'src/models/component.dart';
 export 'src/models/stream_message.dart';
+export 'src/utils/logger.dart';
 export 'src/widgets/gulf_provider.dart';
 export 'src/widgets/gulf_view.dart';
