+++ packages/spikes/gulf_client/lib/src/core/interpreter.dart
@@ -6,8 +6,12 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:flutter/foundation.dart';
+import 'package:logging/logging.dart';
+
 import '../models/component.dart';
 import '../models/stream_message.dart';
+
+final _log = Logger('GulfInterpreter');
 
 /// A client-side interpreter for the GULF Streaming UI Protocol.
 ///
@@ -46,43 +50,65 @@ class GulfInterpreter with ChangeNotifier {
     if (jsonl.isEmpty) {
       return;
     }
+    _log.fine('Processing JSONL message: $jsonl');
     try {
       final json = jsonDecode(jsonl) as Map<String, Object?>;
       final message = GulfStreamMessage.fromJson(json);
+      _log.finer('Parsed message: $message');
       switch (message) {
         case StreamHeader():
+          _log.info('Received StreamHeader: version ${message.version}');
           // Nothing to do for now.
           break;
         case ComponentUpdate():
+          _log.info(
+            'Received ComponentUpdate with ${message.components.length} '
+            'components.',
+          );
           for (final component in message.components) {
+            _log.finer('Updating component: ${component.id}');
             _components[component.id] = component;
           }
           break;
         case DataModelUpdate():
+          _log.info('Received DataModelUpdate at path "${message.path}".');
           _updateDataModel(message.path, message.contents);
           notifyListeners();
           break;
         case BeginRendering():
+          _log.info('Received BeginRendering with root "${message.root}".');
           _rootComponentId = message.root;
           _isReadyToRender = true;
           notifyListeners();
           break;
       }
-    } on UnknownComponentException catch (e) {
+    } on UnknownComponentException catch (e, s) {
       _error = e.toString();
+      _log.severe('Error processing message', e, s);
       notifyListeners();
-      debugPrint('Error: $e');
+    } catch (e, s) {
+      _error = e.toString();
+      _log.severe(
+        'An unexpected error occurred while processing message',
+        e,
+        s,
+      );
+      notifyListeners();
     }
   }
 
   void _updateDataModel(String? path, dynamic contents) {
-    if (path == null || path.isEmpty) {
+    _log.finer('Updating data model at path "$path" with contents: $contents');
+    if (path == null || path.isEmpty || path == '/') {
       if (contents is Map<String, Object?>) {
         _dataModel = contents;
+        _log.finer('Replaced root data model.');
       } else if (contents is Map) {
         _dataModel = contents.cast<String, Object?>();
+        _log.finer('Replaced root data model (after casting).');
       } else {
         _error = 'Data model root must be a JSON object.';
+        _log.severe(_error);
       }
       return;
     }
@@ -146,7 +172,9 @@ class GulfInterpreter with ChangeNotifier {
 
   /// Resolves a data binding path to a value in the data model.
   Object? resolveDataBinding(String path) {
+    _log.finer('Resolving data binding for path: "$path"');
     if (path.isEmpty) {
+      _log.warning('Attempted to resolve empty data binding path.');
       return null;
     }
     final segments = path
@@ -170,9 +198,16 @@ class GulfInterpreter with ChangeNotifier {
           currentValue.containsKey(segment)) {
         currentValue = currentValue[segment];
       } else {
+        _log.warning(
+          'Data binding path segment "$segment" in "$path" does not match '
+          'data structure.',
+        );
         return null; // Path segment doesn't match data structure.
       }
     }
+    _log.finer(
+      'Resolved data binding for path "$path" to value: $currentValue',
+    );
     return currentValue;
   }
