+++ packages/spikes/gulf_client/lib/src/utils/logger.dart
@@ -1,0 +1,17 @@
+// Copyright 2025 The Flutter Authors.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:logging/logging.dart';
+
+/// The logger for the GULF client.
+final gulfLogger = Logger('GULF');
+
+/// Initializes the GULF client logger.
+void initGulfLogger() {
+  Logger.root.level = Level.ALL;
+  Logger.root.onRecord.listen((record) {
+    // ignore: avoid_print
+    print('${record.level.name}: ${record.time}: ${record.message}');
+  });
+}
