+++ packages/spikes/gulf_client/pubspec.yaml
@@ -23,6 +23,7 @@ dependencies:
     sdk: flutter
   freezed_annotation: ^3.1.0
   json_annotation: ^4.9.0
+  logging: ^1.3.0
 
 dev_dependencies:
   build_runner: ^2.8.0
