+++ packages/spikes/gulf_client/test/widgets/gulf_view_test.dart
@@ -94,7 +94,7 @@ void main() {
       await tester.pumpAndSettle();
 
       expect(
-        find.textContaining('Unknown component type: MissingWidget'),
+        find.textContaining('Unknown component: MissingWidget'),
         findsOneWidget,
       );
     });
